+++ src/routes/our-story/+page.svelte
@@ -1,20 +1,9 @@
 <script lang="ts">
-  import { AboutUs, OurStoryScrollable, PageWrapper } from '$lib/components';
+  import { AboutUs, PageWrapper, StorySections } from '$lib/components';
   import { COPY } from '$lib/content';
-  import type { StoryEntry } from '$lib/components/sections/OurStoryScrollable.svelte';
-
-  const ourStoryEntries: Record<string, StoryEntry> = Object.fromEntries(
-    Object.entries(COPY.about.ourStorySections.entries).map(([key, entry]) => [
-      key,
-      {
-        ...entry,
-        bullets: entry.bullets ? [...entry.bullets] : undefined,
-      },
-    ])
-  );
 </script>
 
 <PageWrapper backgroundImage="/images/story-bg.webp" backgroundPosition="center top">
   <AboutUs />
-  <OurStoryScrollable entries={ourStoryEntries} />
+  <StorySections entries={COPY.about.storySections.entries} />
 </PageWrapper>
